clippy module inception: move system::system into system
